docs: automate docs with eslint-doc-generator
