+++ docs/rules/file-extension-in-import.md
@@ -1,6 +1,8 @@
-# n/file-extension-in-import
-> enforce the style of file extensions in `import` declarations
-> - ✒️ The `--fix` option on the [command line](https://eslint.org/docs/user-guide/command-line-interface#fixing-problems) can automatically fix some of the problems reported by this rule.
+# Enforce the style of file extensions in `import` declarations (`n/file-extension-in-import`)
+
+🔧 This rule is automatically fixable by the [`--fix` CLI option](https://eslint.org/docs/latest/user-guide/command-line-interface#--fix).
+
+<!-- end auto-generated rule header -->
 
 We can omit file extensions in `import`/`export` declarations.
