+++ docs/rules/no-extraneous-require.md
@@ -1,6 +1,8 @@
-# n/no-extraneous-require
-> disallow `require()` expressions which import extraneous modules
-> - ⭐️ This rule is included in `plugin:n/recommended` preset.
+# Disallow `require()` expressions which import extraneous modules (`n/no-extraneous-require`)
+
+💼 This rule is enabled in the following configs: ✅ `recommended`, ☑️ `recommended-module`, ✔️ `recommended-script`.
+
+<!-- end auto-generated rule header -->
 
 If a `require()`'s target is extraneous (it's not written in `package.json`), the program works in local, but will not work after dependencies are re-installed. It will cause troubles to your team/contributors.
 This rule disallows `require()` of extraneous modules.
@@ -54,6 +56,10 @@ When an import path does not exist, this rule checks whether or not any of `path
 `tryExtensions` option is the extension list this rule uses at the time.
 
 Default is `[".js", ".json", ".node"]`.
+
+#### convertPath
+
+TODO
 
 ### Shared Settings
