+++ docs/rules/no-process-env.md
@@ -1,5 +1,6 @@
-# n/no-process-env
-> disallow the use of `process.env`
+# Disallow the use of `process.env` (`n/no-process-env`)
+
+<!-- end auto-generated rule header -->
 
 The `process.env` object in Node.js is used to store deployment/configuration parameters. Littering it through out a project could lead to maintenance issues as it's another kind of global dependency. As such, it could lead to merge conflicts in a multi-user setup and deployment issues in a multi-server setup. Instead, one of the best practices is to define all those parameters in a single configuration/settings file which could be accessed throughout the project.
