+++ docs/rules/no-sync.md
@@ -1,5 +1,6 @@
-# n/no-sync
-> disallow synchronous methods
+# Disallow synchronous methods (`n/no-sync`)
+
+<!-- end auto-generated rule header -->
 
 In Node.js, most I/O is done through asynchronous methods. However, there are often synchronous versions of the asynchronous methods. For example, `fs.exists()` and `fs.existsSync()`. In some contexts, using synchronous operations is okay (if, as with ESLint, you are writing a command line utility). However, in other contexts the use of synchronous operations is considered a bad practice that should be avoided. For example, if you are running a high-travel web server on Node.js, you should consider carefully if you want to allow any synchronous operations that could lock up the server.
