+++ docs/rules/prefer-promises/dns.md
@@ -1,5 +1,6 @@
-# n/prefer-promises/dns
-> enforce `require("dns").promises`
+# Enforce `require("dns").promises` (`n/prefer-promises/dns`)
+
+<!-- end auto-generated rule header -->
 
 Since Node.js v11.14.0, `require("dns").promises` API has been stable.
 Promise API and `async`/`await` syntax will make code more readable than callback API.
