+++ lib/rules/no-callback-literal.js
@@ -8,7 +8,7 @@ module.exports = {
     meta: {
         docs: {
             description:
-                "ensure Node.js-style error-first callback pattern is followed",
+                "enforce Node.js-style error-first callback pattern is followed",
             category: "Possible Errors",
             recommended: false,
             url: "https://github.com/weiran-zsd/eslint-plugin-node/blob/HEAD/docs/rules/no-callback-literal.md",
@@ -17,8 +17,9 @@ module.exports = {
         fixable: null,
         schema: [],
         messages: {
-            unexpectedLiteral: "Unexpected literal in error position of callback."
-        }
+            unexpectedLiteral:
+                "Unexpected literal in error position of callback.",
+        },
     },
 
     create(context) {
