+++ lib/rules/no-unsupported-features.js
@@ -1047,14 +1047,15 @@ module.exports = {
                 "disallow unsupported ECMAScript features on the specified version",
             category: "Possible Errors",
             recommended: false,
-            replacedBy: [
-                "n/no-unsupported-features/es-syntax",
-                "n/no-unsupported-features/es-builtins",
-            ],
+
             url: "https://github.com/weiran-zsd/eslint-plugin-node/blob/HEAD/docs/rules/no-unsupported-features.md",
         },
         type: "problem",
         deprecated: true,
+        replacedBy: [
+            "n/no-unsupported-features/es-syntax",
+            "n/no-unsupported-features/es-builtins",
+        ],
         fixable: null,
         schema: [
             {
