+++ lib/rules/process-exit-as-throw.js
@@ -148,7 +148,7 @@ module.exports = {
     meta: {
         docs: {
             description:
-                "make `process.exit()` expressions the same code path as `throw`",
+                "require that `process.exit()` expressions use the same code path as `throw`",
             category: "Possible Errors",
             recommended: true,
             url: "https://github.com/weiran-zsd/eslint-plugin-node/blob/HEAD/docs/rules/process-exit-as-throw.md",
