+++ lib/rules/shebang.js
@@ -69,7 +69,7 @@ function getShebangInfo(sourceCode) {
 module.exports = {
     meta: {
         docs: {
-            description: "suggest correct usage of shebang",
+            description: "require correct usage of shebang",
             category: "Possible Errors",
             recommended: true,
             url: "https://github.com/weiran-zsd/eslint-plugin-node/blob/HEAD/docs/rules/shebang.md",
@@ -89,9 +89,10 @@ module.exports = {
         messages: {
             unexpectedBOM: "This file must not have Unicode BOM.",
             expectedLF: "This file must have Unix linebreaks (LF).",
-            expectedHashbangNode: 'This file needs shebang "#!/usr/bin/env node".',
-            expectedHashbang: "This file needs no shebang."
-        }
+            expectedHashbangNode:
+                'This file needs shebang "#!/usr/bin/env node".',
+            expectedHashbang: "This file needs no shebang.",
+        },
     },
     create(context) {
         const sourceCode = context.getSourceCode()
