+++ package.json
@@ -27,6 +27,7 @@
         "esbuild": "^0.14.39",
         "eslint": "^8.21.0",
         "eslint-config-prettier": "^8.5.0",
+        "eslint-doc-generator": "^0.19.0",
         "eslint-plugin-eslint-plugin": "^5.0.2",
         "eslint-plugin-n": "file:.",
         "fast-glob": "^3.2.11",
@@ -35,6 +36,7 @@
         "import-meta-resolve": "^1.1.1",
         "lint-staged": "^12.4.1",
         "mocha": "^10.0.0",
+        "npm-run-all": "^4.1.5",
         "nyc": "^15.1.0",
         "opener": "^1.5.1",
         "prettier": "^2.7.1",
@@ -47,7 +49,9 @@
         "clean": "rimraf .nyc_output coverage",
         "codecov": "nyc report --reporter text-lcov | codecov --pipe --disable=gcov -t $CODECOV_TOKEN",
         "coverage": "opener ./coverage/lcov-report/index.html",
-        "lint": "eslint lib scripts tests/lib .eslintrc.js",
+        "lint": "npm-run-all \"lint:*\"",
+        "lint:eslint-docs": "npm run update:eslint-docs -- --check",
+        "lint:js": "eslint lib scripts tests/lib .eslintrc.js",
         "new": "node scripts/new-rule",
         "pretest": "npm run -s lint",
         "test": "nyc npm run -s test:_mocha",
@@ -58,7 +62,8 @@
         "postversion": "git push && git push --tags",
         "watch": "npm run test:_mocha -- --watch --growl",
         "format": "prettier --loglevel warn --write \"**/*.{js,css,md}\"",
-        "prepare": "husky install"
+        "prepare": "husky install",
+        "update:eslint-docs": "eslint-doc-generator --config-emoji recommended-module,☑️ --config-emoji recommended-script,✔️ --split-by meta.docs.category"
     },
     "repository": {
         "type": "git",
