+++ scripts/update.js
@@ -4,7 +4,5 @@
  */
 "use strict"
 
-require("./update-docs-header-and-footer")
-require("./update-docs-header-and-footer")
+require("./update-rule-docs")
 require("./update-lib-index")
-require("./update-readme")
